Aula11/requirements.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.4.1

pip~=20.2.4
Markdown~=3.3.3
setuptools~=47.1.0
django-filter~=2.4.0
djangorestframework~=3.12.1